Dont rewrite file when exists.

+++ simple_storage/__init__.py
@@ -85,6 +85,8 @@ class FileStorage():
                                  md5hash,
                                  ext)
         path_to_file = Path(path_to_file)
+        if path_to_file.is_file():
+            return md5hash
         path_to_file.parent.mkdir(parents=True)
         with path_to_file.open('wb') as target_file:
             target_file.write(binary)
